// client/web/src/schema/settings.schema.d.ts
export interface SearchScope {
  /**
   * The human-readable name for this search scope
   */
  name: string;
  /**
   * The query string of this search scope
   */
  value: string;
}
export interface QuickLink {
  /**
   * The human-readable name for this quick link
   */
  name: string;
  /**
   * The URL of this quick link (absolute or relative)
   */
  url: string;
  /**
   * A description for this quick link
   */
  description?: string;
}